fix: 코드/디자인/UX 리뷰 7가지 수정
1. 모바일 편집/삭제 버튼: inline opacity 제거 → CSS로 관리
   (모바일에서 !important 우선순위 확보)
2. 종목 삭제 confirm 추가
3. AnalysisPanel 거래소명: .KS→KRX, .KQ→KOSDAQ, 기타→NASDAQ
4. EditStockModal 통화: 한국주식 ₩, 미국주식 $
5. ToastAlert setTimeout 전체 cleanup (메모리 누수 수정)
6. BuySimulator parseInt→parseFloat (소수점 금액 지원)
7. 뉴스 에러 상태 추가 (영구 로딩 방지)

+++ src/components/portfolio/BuySimulator.tsx
@@ -30,7 +30,7 @@ export default function BuySimulator({
   if (!currentPrice || currentPrice <= 0) return null;
 
   const kr = STOCK_KR[symbol] || symbol;
-  const amountNum = parseInt(amount) || 0;
+  const amountNum = parseFloat(amount) || 0;
   const amountUsd = currency === 'KRW' ? amountNum / usdKrw : amountNum;
   const newShares = Math.floor(amountUsd / currentPrice);
   const actualCost = newShares * currentPrice;
@@ -112,8 +112,8 @@ export default function BuySimulator({
             onClick={() => setAmount(String(p.val))}
             style={{
               flex: 1, padding: '6px 0', borderRadius: 8, fontSize: 12, fontWeight: 600,
-              background: parseInt(amount) === p.val ? 'rgba(49,130,246,0.08)' : '#F2F4F6',
-              color: parseInt(amount) === p.val ? '#3182F6' : '#8B95A1',
+              background: parseFloat(amount) === p.val ? 'rgba(49,130,246,0.08)' : '#F2F4F6',
+              color: parseFloat(amount) === p.val ? '#3182F6' : '#8B95A1',
               border: 'none', cursor: 'pointer',
             }}
           >
